Rename update.ts to updateTags.ts
